requirements.txt
fastapi
uvicorn
jinja2
python-multipart
sqlmodel
python-dotenv
itsdangerous
